+++ src/lib/validation/utils/list.ts
@@ -1,16 +1,16 @@
 import {EnumToken} from "../../ast";
 import type {Token} from "../../../@types";
 
-export function splitTokenList(tokenList: Token[]): Token[][] {
+export function splitTokenList(tokenList: Token[], split: EnumToken[] = [EnumToken.CommaTokenType]): Token[][] {
 
-   return tokenList.reduce( (acc: Token[][], curr: Token): Token[][] => {
+    return tokenList.reduce((acc: Token[][], curr: Token): Token[][] => {
 
         if (curr.typ == EnumToken.CommentTokenType) {
 
             return acc;
         }
 
-        if (curr.typ == EnumToken.CommaTokenType) {
+        if (split.includes(curr.typ)) {
 
             acc.push([]);
         } else {